% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_zenkaku.R
\name{conv_zenkaku_kigou}
\alias{conv_zenkaku_kigou}
\title{Convert zenkaku special characters to hankaku number.}
\usage{
conv_zenkaku_kigou(character_vector)
}
\arguments{
\item{character_vector}{A character vector include zenkaku number character.}
}
\description{
This function convert zenkaku special characters to hankaku characters within string vector.
}
