% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_zenkaku.R
\name{conv_zenkaku_num}
\alias{conv_zenkaku_num}
\title{Convert zenkaku number to hankaku number.}
\usage{
conv_zenkaku_num(character_vector)
}
\arguments{
\item{character_vector}{A character vector include zenkaku number character.}
}
\description{
This function convert zenkaku number to hankaku number within string vector.
}
