% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_zenkaku.R
\name{conv_zenkaku_space}
\alias{conv_zenkaku_space}
\title{Convert zenkaku SPACE to hankaku SPACE.}
\usage{
conv_zenkaku_space(character_vector)
}
\arguments{
\item{character_vector}{A character vector include zenkaku number character.}
}
\description{
This function convert zenkaku SPACE to hankaku SPACE within string vector.
}
