% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input.R
\name{.check}
\alias{.check}
\title{Cheking function for generate errormessage if invalid value given to the function.}
\usage{
.check(expr, errormessage)
}
\arguments{
\item{expr}{An expression to evaluate}

\item{errormessage}{A message to displayed if expr is TRUE}
}
\description{
This function is intended for internal use only
}
