% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_zenkaku.R
\name{.definition_numeric}
\alias{.definition_numeric}
\title{Internal function for convert zenkaku to hankaku}
\usage{
.definition_numeric()
}
\description{
This function will be used with in functions of this package
Not intended to call by package user.
}
