% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_zenkaku.R
\name{.repeat_str_replace_all}
\alias{.repeat_str_replace_all}
\title{Internal function for applying str_replace_all to a single character.}
\usage{
.repeat_str_replace_all(a_character, froms, tos)
}
\arguments{
\item{a_character}{Single character.}

\item{froms}{vector containing argument for pattern in str_replace_all.}

\item{tos}{vector containing argument for replace in str_replace_all.}
}
\description{
Internal function for applying str_replace_all to a single character.
}
