% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input.R
\name{.warn}
\alias{.warn}
\title{Warning function for generate warningmessage if warning message need to be printed}
\usage{
.warn(expr, warningmessage)
}
\arguments{
\item{expr}{An exporession to evaluate}

\item{warningmessage}{A message to displayed if expr is TRUE}
}
\description{
This function is intended for internal use only.
}
