% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gengo.R
\name{heisei_to_seireki}
\alias{heisei_to_seireki}
\title{Convert "Heisei" to Jurian years (Seireki).}
\usage{
heisei_to_seireki(integer_vector)
}
\arguments{
\item{integer_vector}{Integer vector with minimum value 1.}
}
\description{
Convert "Heisei" to Jurian years (Seireki).
}
