% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation].R
\name{remove_spaces}
\alias{remove_spaces}
\title{Function for remove white spaces.}
\usage{
remove_spaces(character_vector, zenkaku = TRUE)
}
\arguments{
\item{character_vector}{Character vector to remove spaces}

\item{zenkaku}{Boolean. TRUE(default) to remove zenkaku and hankaku spaces.}
}
\description{
Function for remove white spaces.
}
