% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_zenkaku.R
\name{str_replace_all_repeats}
\alias{str_replace_all_repeats}
\title{Function for apply multiply str_reoplace_all condition to character vector}
\usage{
str_replace_all_repeats(character_vector, froms, tos)
}
\arguments{
\item{character_vector}{Vector of character.}

\item{froms}{Vector containing argument for pattern in str_replace_all}

\item{tos}{Vector containing argument for replace in str_replace_all}
}
\description{
This is wrapper for str_replace_all to apply multiple condition to the character vector
}
